// src/domain/_layer/infrastructure/service/notification/notification-types.enum.ts
export enum NotificationType {
  CONTACT_SUPPORT = 'contact-support',
  USER_ALREADY_REGISTERED = 'user-already-registered',
  USER_RECOVER_PASSWORD = 'user-recover-password',
  USER_REACTIVE_ACCESS = 'reactive-access',
  QUERY_AUTO_REPROCESS_FAILED = 'query-auto-reprocess-failed',
  QUERY_AUTO_REPROCESS_SUCCESS = 'query-auto-reprocess-success',
  QUERY_SUCCESS = 'query-success',
  QUERY_FAIL = 'query-fail',
  QUERY_ALERT = 'alert-on-query-make',
  QUERY_REVISION_PLAN = 'alert-on-query-revision-plan',
  QUERY_FINE = 'alert-on-query-fine',
  QUERY_FIPE_PRICE = 'alert-on-query-fipe-history',
}